rename the overlapped name
